Remove site.name var from index page
The page title should be used by default.
The site.name variable is thus superfluous, but also causing an odd issue of '{} | ...' being rendered as the title.
This is presumably something to do with how Jekyll, or one of the plug-ins in use, combines values to generate the site name.
This GH issue suggests site.title is preferable in any case:
https://github.com/jekyll/jekyll-seo-tag/issues/238