Add links between nodes
The collector-volunteer arrow shows the direction of the dependency rather than the communication.
This is because I want the the consuming system on the opposite side to the dependencies of the Spartakus volunteer.
I might revisit this, as it doesn't reflect the active/passive sides of the comms and direction of information transfer.