Remove blog content & reorder sections for narrative flow
Beyond the introduction, the existing content veered into the matter for the post on Core v1 vs. Core v2.
The post on the operator should be focused on how the operator fits in and what changed from v1 to v2, but without worrying over much about the whys and wherefores behind that.
The existing structure furthermore went into too much detail, taking content from later sections and detailing it early on.
Instead, I'd like to have a flow that discusses operators generally, the role of them for Core, and then specifics and interesting diversions from norms in Core v2.